refactoring addOneMin
